test(view-05): fold smoke into modeling-shell spec to share single login
Standalone products-view-05.spec.ts pushed cumulative login count past
the 5/IP/15min auth rate-limiter ceiling once retries kicked in (the
multi-tenant-isolation spec already burns 3 raw API logins). Lessons.md
§10 explicitly recommends consolidating into modeling-shell when the
suite would exceed ~4 logins; the VIEW-05 surface lives at /products
and only needs one shared login to assert.

Refs #411
