--- apps/admin/e2e/products-view-05.spec.ts
@@ -1,93 +1,0 @@
-import { expect, test } from '@playwright/test';
-
-import { loginAsAdmin } from './helpers/auth';
-
-test.use({ locale: 'pl-PL' });
-
-/**
- * VIEW-05 (#411) — pixel-perfect delta-alignment of the products list.
- * Asserts the new header copy, SavedViewsRail, FilterPill toolbar,
- * VariantsToggle segmented control, ProductsGrid 12-column layout, and
- * BulkBar sticky bottom bar with toast placeholders.
- *
- * Single test scenario with sequential sections so the suite costs only
- * one login against the 5/IP/15min auth rate-limiter (lesson from
- * `agent/lessons.md` §10 — UI-03 marathon flak). Axe-core a11y scan
- * is deferred to follow-up VIEW-05.0 once `@axe-core/playwright` lands
- * in deps; manual Lighthouse run substitutes for the MVP merge.
- */
-/**
- * i18n note: assertions match BOTH PL and EN copy because i18next
- * `LanguageDetector` reads Accept-Language from the browser context;
- * Playwright Chromium defaults to `en-US` regardless of the
- * `locale: 'pl-PL'` project knob, so the spec must tolerate either
- * locale to stay green in CI.
- */
-test('VIEW-05 Products · Lista — pixel-perfect smoke', async ({ page }) => {
-  await loginAsAdmin(page);
-  await page.goto('/products');
-  await expect(page).toHaveURL(/\/products$/);
-
-  // Section 1 — header + toolbar + grid render pixel-perfect.
-  await expect(page.getByText(/Workspace · (katalog|catalog)/, { exact: true })).toBeVisible();
-  await expect(page.getByRole('heading', { level: 1, name: /Produkty|Products/ })).toBeVisible();
-  await expect(
-    page.getByRole('searchbox', { name: /Szukaj produktów|Search products/i }),
-  ).toBeVisible();
-  await expect(page.getByRole('button', { name: /^(Marka|Brand)$/ })).toBeVisible();
-  await expect(page.getByRole('button', { name: /^(Rodzina|Family)$/ })).toBeVisible();
-  const channelBtn = page.getByRole('button', { name: /^(Kanał|Channel)$/ });
-  await expect(channelBtn).toBeVisible();
-  await expect(page.getByRole('button', { name: /^Status$/ })).toBeVisible();
-  await expect(page.getByRole('button', { name: /^(Płasko|Flat)$/ })).toBeVisible();
-  const treeBtn = page.getByRole('button', { name: /^(Drzewo|Tree)$/ });
-  await expect(treeBtn).toBeVisible();
-  await expect(page.getByRole('link', { name: /(Nowy produkt|New product)/i })).toBeVisible();
-  await expect(page.getByTestId('products-grid')).toBeVisible();
-
-  // Section 2 — channel pill click surfaces toast (epic 0.6 placeholder).
-  await channelBtn.click();
-  await page.getByRole('menuitem', { name: 'Shopify' }).click();
-  await expect(
-    page
-      .getByRole('status')
-      .filter({
-        hasText: /(Filtr per kanał czeka na epik 0\.6|Per-channel filter waits for epic 0\.6)/,
-      })
-      .first(),
-  ).toBeVisible();
-
-  // Section 3 — variants segmented toggle flips active state on click.
-  const flatBtn = page.getByRole('button', { name: /^(Płasko|Flat)$/ });
-  await treeBtn.click();
-  await expect(treeBtn).toHaveAttribute('aria-pressed', 'true');
-  await expect(flatBtn).toHaveAttribute('aria-pressed', 'false');
-  await flatBtn.click();
-  await expect(flatBtn).toHaveAttribute('aria-pressed', 'true');
-  await expect(treeBtn).toHaveAttribute('aria-pressed', 'false');
-
-  // Section 4 — selecting a row reveals BulkBar with placeholder toasts.
-  const grid = page.getByTestId('products-grid');
-  const firstRow = grid.locator('[data-testid^="products-grid-row-"]').first();
-  const rowCount = await firstRow.count();
-  if (rowCount === 0) {
-    test.info().annotations.push({
-      type: 'note',
-      description: 'no products in fixtures — BulkBar section skipped',
-    });
-    return;
-  }
-  await firstRow.getByRole('checkbox').first().check();
-  const bulkBar = page.getByTestId('bulk-bar');
-  await expect(bulkBar).toBeVisible();
-  await expect(bulkBar).toContainText(/(zaznaczonych produktów|selected products)/);
-  await bulkBar.getByRole('button', { name: /(Edytuj atrybut|Edit attribute)/ }).click();
-  await expect(
-    page
-      .getByRole('status')
-      .filter({ hasText: /(W przygotowaniu|In progress) — VIEW-05\.2/ })
-      .first(),
-  ).toBeVisible();
-  await bulkBar.getByRole('button', { name: /^(Wyczyść|Clear)$/ }).click();
-  await expect(bulkBar).not.toBeVisible();
-});
